Move tests to existing test file
