+++ tests/epyccel/test_epyccel_lists.py
@@ -690,3 +690,55 @@ def test_mutable_multi_level_indexing(stc_language):
 
     epyc_f = epyccel(f, language=stc_language)
     assert f() == epyc_f()
+
+def test_homogenous_list_int_copy(language):
+    def homogeneous_list_int():
+        return list([1, 2, 3, 4])
+    f1 = homogeneous_list_int
+    f2 = epyccel( f1 , language=language)
+
+    python_out = f1()
+    pyccel_out = f2()
+    print(pyccel_out)
+    print(python_out)
+
+    assert python_out == pyccel_out
+
+def test_homogenous_list_bool_copy(language):
+    def homogeneous_list_bool():
+        return list([True, False, True, False])
+    f1 = homogeneous_list_bool
+    f2 = epyccel( f1 , language=language)
+
+    python_out = f1()
+    pyccel_out = f2()
+    print(pyccel_out)
+    print(python_out)
+
+    assert python_out == pyccel_out
+
+def test_homogenous_list_float_copy(language):
+    def homogeneous_list_int():
+        return list([1.0, 2.0, 3.0, 4.0])
+    f1 = homogeneous_list_float
+    f2 = epyccel( f1 , language=language)
+
+    python_out = f1()
+    pyccel_out = f2()
+    print(pyccel_out)
+    print(python_out)
+
+    assert python_out == pyccel_out
+
+def test_homogenous_list_int_tuplecopy(language):
+    def homogeneous_list_int_tuple():
+        return list((1, 2, 3, 4))
+    f1 = homogeneous_list_int_tuple
+    f2 = epyccel( f1 , language=language)
+
+    python_out = f1()
+    pyccel_out = f2()
+    print(pyccel_out)
+    print(python_out)
+
+    assert python_out == pyccel_out
